add GUI abstract class with setup methods
The setup methods are for JLabel, JButton, JPanel
